MM-11477 Add ability to provide custom thunk middleware
