+++ src/store/configureStore.dev.js
@@ -4,11 +4,8 @@
 
 import {createStore} from 'redux';
 import devTools from 'remote-redux-devtools';
-import thunk from 'redux-thunk';
-import {REHYDRATE} from 'redux-persist/constants';
 import {createOfflineReducer, networkStatusChangedAction, offlineCompose} from 'redux-offline';
 import defaultOfflineConfig from 'redux-offline/lib/defaults';
-import createActionBuffer from 'redux-action-buffer';
 import reducerRegistry from 'store/reducer_registry';
 import {Client4} from 'client';
 
@@ -28,31 +25,12 @@ const devToolsEnhancer = (
 import serviceReducer from 'reducers';
 import deepFreezeAndThrowOnMutation from 'utils/deep_freeze';
 import initialState from './initial_state';
-import {defaultOptions, offlineConfig, createReducer} from './helpers';
+import {offlineConfig, createReducer} from './helpers';
+import {createMiddleware} from './middleware';
 
-/***
-clientOptions object - This param allows users to configure the store from the client side.
-It has two properties currently:
-enableBuffer - bool - default = true - If true the store will buffer all actions until offline state rehydration occurs.
-additionalMiddleware - func | array - Allows for single or multiple additional middleware functions to be passed in from the client side.
-***/
-export default function configureServiceStore(preloadedState, appReducer, userOfflineConfig, getAppReducer, clientOptions = {}) {
+export default function configureServiceStore(preloadedState, appReducer, userOfflineConfig, getAppReducer, clientOptions) {
     const baseOfflineConfig = Object.assign({}, defaultOfflineConfig, offlineConfig, userOfflineConfig);
-    const options = Object.assign({}, defaultOptions, clientOptions);
     const baseState = Object.assign({}, initialState, preloadedState);
-
-    const {additionalMiddleware, enableBuffer} = options;
-
-    let clientSideMiddleware = additionalMiddleware;
-
-    if (typeof clientSideMiddleware === 'function') {
-        clientSideMiddleware = [clientSideMiddleware];
-    }
-
-    const middleware = [thunk, ...clientSideMiddleware];
-    if (enableBuffer) {
-        middleware.push(createActionBuffer(REHYDRATE));
-    }
 
     const loadReduxDevtools = process.env.NODE_ENV !== 'test'; //eslint-disable-line no-process-env
 
@@ -61,7 +39,7 @@ export default function configureServiceStore(preloadedState, appReducer, userOf
         baseState,
         // eslint-disable-line - offlineCompose(config)(middleware, other funcs)
         offlineCompose(baseOfflineConfig)(
-            middleware,
+            createMiddleware(clientOptions),
             loadReduxDevtools ? [devToolsEnhancer()] : []
         )
     );
